requirements.txt
streamlit>=1.0.0
langchain==0.1.16
pandas>=2.0.0
openpyxl>=3.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
faiss-cpu>=1.7.0
sentence-transformers>=2.0.0
ollama>=0.1.0
python-dotenv>=1.0.0